requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.0
pandas==2.2.3
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.7.2
jinja2==3.1.4
python-multipart==0.0.9
openpyxl==3.1.5

